regular update

+++ app/src/main/java/com/example/weatherapp/Models/Coord.java
@@ -1,26 +1,26 @@
 package com.example.weatherapp.Models;
 
 public class Coord {
-    private float lon,lat;
+    private double lon,lat;
 
-    public Coord(float lon, float lat) {
+    public Coord(double lon, double lat) {
         this.lon = lon;
         this.lat = lat;
     }
 
-    public float getLon() {
+    public double getLon() {
         return lon;
     }
 
-    public void setLon(float lon) {
+    public void setLon(double lon) {
         this.lon = lon;
     }
 
-    public float getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(float lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
 }
